move runDB_declarative to db namespace
