+++ htsohm/db/__init__.py
@@ -1,0 +1,1 @@
+from htsohm.db.runDB_declarative import *
